test(read amplification): cover multishard_combining_reader issue
Cover issue: https://github.com/scylladb/scylla/issues/8161
Commit: https://github.com/scylladb/scylla/commit/bc1fcd3db20eb957524387214617b613f1cab3e7

The multishard combining reader currently assumes that all shards have
data for the read range. This however is not always true and in extreme
cases (like reading a single token) it can lead to huge read
amplification.
After this commit, the multishard reader will only read from shards that
have data relevant to the read range, both in the case of normal reads
and also for read-ahead.
Task: https://trello.com/c/agJQnBCf/3486-8161-the-multishard-combining-reader-is-responsible-for-
merging-data-from-multiple-cores-when-a-range-scan-runs-a-bug-that-is-tr