test(manager): tests regarding the new ignore-down-hosts parameter
New in manager 2.6
Added a new argument to the repair task api: ignore-down-hosts
The argument allow repair tasks to run even if some of the nodes are DN.

In this PR there are also two tests of the ignore-down-hosts argument:

The first test starts a fully functioning cluster and runs a repair task with the
ignore-down-hosts, and the task is expected to succeed

The second test start a cluster and shuts down one of the nodes. Afterwards, the
test starts two repairs: one without ignore-down-hosts and one with. The task expects
the repair without the argument to fail and repair with the argument to succeed.