alternator_ttl_test.py: reduce code duplication
The two Alternator TTL tests were more-or-less identical with one
difference - one ran on a 5-node cluster, and the other ran on a 5-node
cluster with one node shut down.

As suggested by @Orenef11, we can write the test code just once, and
then use pytest's "parametrize" feature to run the same test twice -
once with all nodes up, and once with one node brought down.

This patch also switches to a nicer idiom for naming two of the nodes:
   node1, *_, node5 = self.cluster.nodelist()

Signed-off-by: Nadav Har'El <nyh@scylladb.com>