cql_system_clients_test: update require marked from #9646 to #9216
These tests were originally decorated with `@skip('#6946')`
in d4f134e7e and 01cc6b08f.

That was converted to pytest.mark.require in 87ae715c8
as part of the pytest conversion.

However, scylladb/scylla#6946 was closed in favor of
scylladb/scylla#9216, so update the require marked
to the proper issue since none of these test
can pass now.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>