materialized_views_test: TestMaterializedViews: delete test_interrupt_build_process_* cases
They all moved to TestInterruptBuildProcess in 7e7b51859
and are executed one at a time using a Lock..
Also, all the relevant fixes were applied there.

There is no need to repeat them in this class,
especially that not all pass after the transition to pytest.
For example:
https://jenkins.scylladb.com/view/master/job/scylla-master/job/dtest-release/1069/testReport/materialized_views_test/TestMaterializedViews/test_interrupt_build_process_and_resharding_half_to_low/
can't pass without 49d2953d3.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>