listsnapshots_test: compare_filesize_and_output: allow size to be between the logical size and allocated size
With https://github.com/scylladb/scylla/issues/9898 fixed,
scylla will return the on-disk allocated size for the files
in the snapshot.  Fot this test, it typically returns 40 KB instead
of 4.99 KB, depending on the filesystem.

Change the test to accomodate for both, and count both
the logical size fo the files and the on-disk allocated size
using os.lstat and test the listsnpashots returns a number in between.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>