cql_additional_tests.py: Add test_indexed_statement_concurrency_limit
This commit adds a new test intended to test scylla behaviour post
3e3003fcc14ca2ca5212c22ca6757c3ba9b5239c, which introduces concurrency
limit for paged indexed queries. The test asks for a page of 1024 rows,
in a table where each row is around 4KiB in size, tripping scylla 1MiB
limit at around 511 rows and causing it to return first page with that
amount of rows, instead of defined page size.