commitlog_test: check_commitlog_size: check actual_space_limit only when not allowing errors
When the node runs in parallel to du there's a chance
that space will be over-accounted if files are deleted
and added during du's scan of the directory, as du
may account space both for files that are deleted
during the scan plus files added.

Therefore, test `dir_size > actual_space_limit`
only if `not allow_errors`, when called
when the node is stopped.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>