materialized_views_test: do_reasharding_test: expand expected Reshard log pattern
As reported in https://github.com/scylladb/scylla/issues/9946
the test fails occasionally when not finding the exact
`Reshard ks` message, but otherwise, the log indicates that resharding
was considered and maybe was not required.

For the purpose of this test, we just want to inject a nemesis
by stopping the node abruptively during resharding and then
we test the data and the materialoized view throughly after
the ndoes restart, so if resharding was needed and didn't happen
verifying the data must catch it.

This change makes the log pattern more liberate and will
catch the first `Resharding XMB for ks.t` message before
stopping the node, to fix the test flakiness.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>