dtest_setup: __filter_errors: filter rpc::unknown_verb_error
We hit it from time to time when nodes start/stop
and it is assumed to be benign in dtests if
it doesn't cause any other malfunction.

As seen in e.g. https://jenkins.scylladb.com/job/scylla-master/job/dtest-release/1103/testReport/bootstrap_test/TestBootstrap/test_killed_wiped_node_cannot_join/

If it does, the test must test the high-level
functionality to verify the system correctness.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>