wide_rows_test.py: compare sstable basename rather than full path
Prepare for a corresponding change in scylla
that will keep only the sstable basename in the system.large*
tables.

SSTables may be created in one path, e.g. under the staging/
sub-directory, where the large data records are created,
and then moved to a different directory (base directory
in case of staging, or to quarantine after scrub), but
the large data records apply to the sstable whereever it
is stored, and that is uniquely identifed by the basename
(that includes the generation number and the sstable format)

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>