update_cluster_layout_tests: test_simple_kill_new_node_while_bootstrapping_with_parallel_writes: kill the bootstrapping node during bootstrap
Waiting for the nodetool status for the joining node to change
proves to be too long, causing the test to kill the node
after bootstrapping completed.  This, in turn generated
Unavailable errors for the SELECT query since the storage_proxy
thinks the new node owns the token moved to it.

For example, as reproduce with storage_proxy trace logging
for writes after the node was killed:
```
TRACE 2022-02-17 15:00:55,695 [shard 0] storage_proxy - creating write handler for token: 4319725971821055942 natural: {127.0.98.2, 127.0.98.4} pending: {}
```

This change waits for an early indication of the streaming plan

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>