alternator_ttl_test: configure Alternator TTL to finish quickly
The tests in alternator_ttl_test were written to assume that expired
items get deleted within a few seconds, using a timeout of 1 minute.
However, recently the default expiration period was increase to 24
hours, so we need to fix the test to reduce this period (we reduce
it to one second) if we want the test to finish in a reasonable
timeframe and not fail like it does now.

Fixes https://github.com/scylladb/scylla/issues/10141

Signed-off-by: Nadav Har'El <nyh@scylladb.com>