fix(manager backup): alter the way we filter manager created snapshots
Since the cluster under test is also used as the backend of the manager server,
it sometimes creates snapshots of its own that are not related to the backups
or to the test. Previously we would just filter out 'scheduler_task' snapshots,
but it seems that there other types now. Instead of filtering them out, we decided
to filter out all snapshtos of the scylla_manager keyspace, since they are not a
part of the test.