tests: Fix the 'JOINING: Starting to bootstrap' message
There's an ongoing work to improve operation mode handling in scylla. Part
of it includes fixing the "Starting to bootstrap" messages -- they will lose
the "JOINING:" prefix. This set prepares dtest for the change by checking
only the message payload, so it's compatible with both -- current scylla and
soon-to-be-patched one.

Tested with bootstrap_test full run on both scyllas.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>