repair additional tests: trigger off-strategy compaction using rest api
This test currently needs to wait for 5 minutes until
off-strategy compaction is auto-triggered on timeout after repair.
Explicitly use the rest api to trigger off-strategy compaction
immediately.

Note that this doesn't invalidate the test
as off-strategy compaction will happen when triggered
over the rest api only if it was required.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>