refactor(manager task): Fetching the task arguments/properties from sctool info
New in manager 3.0
Previously, the manager printed the arguments used in the task
in 'task progress' and 'task list' command.
However, since manager 3.0, the arguments will not be printed in
the output of those command, but instead it will be printed in the
output of the 'sctool info' command.
I altered the tests that relied on the old way the arguments were
printed to fetch them from 'sctool info' instead.