fix(test_create_scheduled_task_while_suspended): Changed expected result
Previously, while the manager was suspended, you could schedule a task to run in
the future as long as the run time was more than 8 hours in the future.
As of 3.0, this is no longer possible. One cannot schedule a task in any time while
the manager is suspended.
As such, I altered the test to expect this failure.
Also lowered node count from 3 to 2, since there's no need for 3