fix(test_update_backup_parameters): Added keyspace and bucket creation
test_update_backup_parameters is a test that creates a backup, updates its params
and then makes sure that the params were indeed updated.
Previously, the update command would just change the params without checking any of
them, so in the test we just used an imaginary keyspace and bucket for the update,
and the update worked.
Since manager 3.0, one cannot update a backup task to use a keyspace or bucket that
does not exist, since the manager checks their existance upon the execution of the
update command.
As such, in test_update_backup_parameters I created both the new keyspace and the
new backup before the update command execution, so that the update command would
succeed and the test can continue