update_cluster_layout_tests.py: Do not check UL when decommission fails
In test_simple_kill_remained_node_while_decommissioning, when node1
stops, the decommission of node2 will fail. With recent scylla having
node_cmd_ops support, the cluster will revert to the state as if the
decommission did not happen in case decommission fails. So we should not
assume and check node2 is still in UL state.

Fixes: https://github.com/scylladb/scylla/issues/10664