compaction_additional_test: `run_sstablemetadata` on less sstables
the verification of this test is reading all the sstables, and it
can take upto 40min, reading only the 10 first and the 10 last
sstables, achive the same goal while taking ~20sec.

also checking the sstable with `time.time()` doesn't mean anything
it's not even in the same scale, one is in secoands on in microseconds.
also the timestamps are generated, and not real clock timestamps.

so we can assert to the exact expected min/max timestamps values.