lwt: update cql_conditional_tests output for new batch formatting
Scylla LWT batch output formatting is different from Cassandra,
and recently it changed one more time.

We always return the previous value of a row in an LWT update,
regardless of whether the update succeeds or not. We always return one
row per batch statement. If a statement in a batch returns a static
row, we return the previous value of the static row.

This is exactly what has changed recently (for a static row update,
we used to return the first regular row of the partition, now we return
the static row instead).

Update the test accordingly.