partitioner_tests: ignore cdc error in unsupported upgrade path
There is a bug in Scylla which causes nodes to enable features
prematurely. The bug masks a problem inside this dtest. When we fix the
bug (see scylladb/scylladb#11225), the problem is unmasked and the test
starts to fail. The failure is caused by an unsupported scenario in
Scylla, where we upgrade from a very old version (the test does an even
weirder thing - it starts Scylla using work directories created by
Cassandra). The cluster boots, but there's an ERROR in the log. Since we
expect it, ignore it.