limits_test.py: increase query_tombstone_page_limit
For test_max_cells and test_max_columns_and_query_parameters. These two
tests generate enough tombstones in `system_schema.columns` that the
driver hits the tombstone limit when trying to query the schema tables
after a schema change. The real fix would be for the driver to use a
paged query, but that is not a trivial fix. In the meanwhile let these
tests pass by increasing the query tombstone page limit such that they
don't hit the limit.

Fixes: #2825