raft, fix for limits_test.py::TestLimits::test_max_columns_and_query_parameters
Related scylla PR scylladb/scylladb#11318.

The limit value for Raft is half as much
to protect against paddings that the
commit log can add for each mutation, as well as
against different commit log limits on different nodes.

With raft the exception occurs at a different time,
while saving the raft command, this is
the reason for a slightly different error message.