Make wide_rows_test pass with Raft based schema changes
The test uses a cluster with 2 nodes and tries to
create a table while one of them is down. Extend the cluster
size so that even with 1 node down there is a majority,
the cluster size is not relevant to what the test is doing
(checking the wide rows information availability on a single node).

With this patch, the test should work equally well with or without
Raft based schema changes.