test(distributed_queries): add test for count(*)
This commit adds a test for the distributed execution of
count(*) queries. The test is meant to check if toggling
the feature flag "enable_parallelized_aggregation"
improves the execution time of the query. We also assert
that the results are the same and that after the toggle
the query request is forwarded to other coordinator nodes.
No assumption is made for the ratio between the distributed
and non-distributed execution, other than that the
first one should be faster.