test(limits): per partition rate limit tests
It is now possible to limit read rates and writes rates into a partition
with a new WITH per_partition_rate_limit clause for the
CREATE TABLE and ALTER TABLE statements.
This is useful to prevent hot-partition problems when high rate
reads or writes are bogus (for example, arriving from spam bots)

This commit adds test for this feature.

refs:
https://github.com/scylladb/scylla/commit/dab56b82fae5e36f7aa2ca6700d8cfa5baa2b515
https://github.com/scylladb/qa-tasks/issues/806