dtest_setup: init_default_config: always update timeout config
Currently, if self.cluster_options is already set
we don't update the timeout config.

Most if not all of the resharding tests fail
due to timeout of `SELECT count(*)` query
with scylla in debug mode, since the range_request_timeout_in_ms
default value of 10 seconds is too small
for debug mode.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>