distributed_queries_test: reduce number of ops in debug mode
Since the test times out consistently after 2700 seconds.

While at it, cut the number of ops in non-debug mode in half
since it needlessly takes a long time to populate the data using stress
and we should aim to keep the test running in the 3-5 minutes time
frame in release/dev modes.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>