update_cluster_layout_tests: harden test_simple_add_two_nodes_in_parallel
Currently the test fails from time to time when node3
starts too late, relative to node2. and it doesn't notices
that node2 is booting, and node3 starts normally.

The relative timing of starting node2 and node3 is
hard to adjust exactly so that node3 will notice node2
at just the rigt state, and refuse to start up.

This change introduces some randomization to timing
node3's start, and allows great variation in monitoring
node3 log.

It will pass if node3 either failed early, detecting node3
`has gossip status=UNKNOWN`, or otherwise, it will look for
`sleep 1 second and check again` and will allow node1
to bootstrap normally eventually if it was started late.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>