test_max_cells fix
There were two bugs in the _do_test_max_cell_count
function: session.execute was called for i == 0, i.e.
before batch_size was reached, and it was not called
for the last batch. Both problems are fixed by the
modified if condition.

The above means the only batch that was executed before
was the first one, consisting of a single insert command.
To be able to execute multi-insert batches we needed
to adjust various limits. First, MAX_CELLS_BATCH_SIZE
is way too large, reduced it to 50. Second, we need
more memory to accommodate larger batches, so the
settings for --memory and --smp. The same is for options batch_size_warn_threshold_in_kb and
batch_size_fail_threshold_in_kb also.

The change of commitlog_segment_size_in_mb is needed
for raft, otherwise CREATE TABLE hits the
raft command size limit.