replace_address_test: do not rely on host_id collision
... and don't bother to restart the replaced node
until scylladb/scylladb#5523 is fixed, there is no
point in restrating the replaced node, as it will be able
to steal back its replaced tokens.

The current method of checking for the host_id collision
message on the other nodes doesn't make sense
since it will match the first collision messages that
are already in the log when node4 replaced node3,
while restarting node3 will generate the opposite collision
messages when node3 is wrongfully able to replace node4 when restarted.

Plus, a change on scylla side will make the host_id
unique so there will be no host_id collission at all.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>