resharding_test: _verify_row_number: adjust query timeout for count(*)
Commit d0c57233aa5120edd0db106d15ecb073d177b6b5
changed _verify_row_number to use the default timeout
(which is 90 seconds in debug mode), but this we still see
a lot of timeouts of this query due to extremely slow
execution of the distributed count(*) query
(takes over 13 minutes on my machine).

Use self.count_request_timeout as the cql (client)
timeout which is 1200 seconds (20 minutes) in debug mode.
And update the server `range_request_timeout_in_ms` respectively.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>