compaction_additional_test.py: stop printing all the query results
on cases of failures `test_enable_disable_optimized_query_for_twcs`
was print the whole content of the DB into the log, and into
the assert statement, which in turn was filling jenkins output and
causing failures to plugins reading the output.

switch from asserting the equality of those results, to using
DeepDiff, that can show only the difference, when there is such
difference.

Fixes: #2936