fix: create table with SizeTierd compaction strategy explicitly
In the test migration_test.TestTTLWithMigrate.test_big_table_with_ttls
we are trying to migrate a table to Cassandra, this works well when we
are on the OSS version since the default compaction strategy for newly
created tables is SizeTiered. However in the enterprise branches it the
default compaction strategy is ICS which Cassandra knows nothing about
causing it to fail with:
`Unable to find compaction strategy class
'org.apache.cassandra.db.compaction.IncrementalCompactionStrategy'`

This commit simply adds the compaction strategy explicitly so it will
also work with enterprise versions or when the default strategy will
change to one that has no Cassandra counterpart.

Tests:
Ran the test before and after the change on OSS and Enterprise version
of scylla and made sure that before the change only OSS passes and after
the change both version passes.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>