system_tables_test: use node.memory() to verify memory.total
* do not pass "--memory 1G" in
  TestRuntimeInfoTable.test_memtable_metrics
* use node.memory() to verify memory.total

before this change, we compare 1024^3 with the reported total memory.
but 1024^3 is but a magic number despite that we pass "1G" via jvm_args
to prepare the 1-node cluster. the connection between "1024 * 1024 *
1024" and jvm_args is not very obvious.

after this change, the reported total memory size is compared with
the one returned by `ScyllaNode.memory()`. and the default settings
for resource allocation are used. so:

1. we don't need to convert "1G" to a size in bytes. simpler this way.
2. we don't need to specify "1G" in the command line, less magic
   numbers!

this change is inspired by
https://github.com/scylladb/scylladb/issues/12522

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>