*_test: mark tests using 1 node with single_node
the document marker of "single_node" in `pytest.ini` reads:

> test that are mark like this, should be using only one node, and
> should boot much quicker

so in this change, all tests using a cluster prepared using

```py
self.prepare_cluster(nodes=1)
```

or
```py
self.cluster.populate(1)
```

are marked with `@pytest.mark.single_node` mark. this would allow
us to customize the fixture setting for the test. for instance,
to skip the gossip to settle and hence to prepare the cluster
faster, so we can perform the tests at a quicker pace.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>