raft: disable consistent_cluster_management if it wasn't passed to dtest
In https://github.com/scylladb/scylladb/pull/12585 we enable
`consistent_cluster_management` in conf/scylla.yaml by default. But in
dtest we want `consistent_cluster_management` to be enabled only if it's
explicitly requested, because we have pipelines that are supposed to run
without Raft. The pipelines that are supposed to run with Raft pass
`--consistent-cluster-management` to dtest so it will explicitly enable
the flag in Scylla config.

Set `consistent_cluster_management` to `False` if it's not requested by
the user of dtest.