test_cdc_large_values: increase blocked_reactor_notify_ms
We see spurious reactor stalls due to the large blobs
and large queries that are unrelated to the stalls checked
when using prepare_statements.

Since we don't have a mechanism in place to tolerate
some stalls and fail on others, increase the threshold to 100ms
in non-debug modes to fail only excessive stalls.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>