system_tables_test: be more tolerant when checking total_memory
because of the additional reserved memory for UDF, the memory
allocated for each shard might not be aligned to 2MB, which is
required by Seastar. and Seastar always aligns the memory allocated
to each shard to 2MB. so, for instance, if we launch a Seastar
application with `--smp 2`, and `reserve_additional_memory` of
`50MB`, `total_memory` of `1GiB`, each shard would get

(1024 - 50) / 2 = 487 MiB

but 487 MiB is not a multiple of 2 MiB, so Seastar just rounds it
down to 486 MiB. this sums up to, (486 * 2) MiB, i.e., 972 MiB.
Hence the total memory managed by Seastar allocator would be
972 MiB. if we take the additional reserved memory for non-seastar
allocator. the total memory would be (972 + 50) = 1022 MiB, which
is 2 MiB less than the specified total memory size of 1024 MiB.

in this change, we take the alignment into consideration. in the
worst case the difference would be (nproc * alignment). so, instead
of using "assert(reported_total_size == configured_size)", we use
"assert(configured_size - nproc * alignment <
        reported_total_size <=
        configured_size)"

in hope to avoid hardwiring the test to the size of memory preserved
for UDF.

Fixes https://github.com/scylladb/scylladb/issues/12522
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>