ttl_test: remove outdated test
The test 'test_remove_column_ttl_with_default_ttl' was an old Cassandra
test from 2014, that checked that "USING TTL 0" doesn't work as documented
when the table has a default TTL - it created a modification with the
default TTL instead of an unexpiring item.

However, already in 2016 Cassandra decided this was bug, not a feature
(see https://issues.apache.org/jira/browse/CASSANDRA-11207) and the
fix reached Cassandra 3.6. We now plan to make the same fix also in
Scylla - see https://github.com/scylladb/scylladb/pull/13079.

So let's drop the incorrect test that will fail with fixed Scylla.
The scylla.git's cql-pytest tests already includes newer tests for
this scenario (including one translated from Cassandra), so it's fine
to remove dtest with writing a new one to replace it.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>