cdc_test: use correct CL to query CDC streams tables (this time for real)
In 0557e935cea4ada969d51ff4d8e4ad547f712b68 the code was modified to
query CDC_STREAMS_TABLE and CDC_TIMESTAMPS_TABLE with QUORUM instead of
ONE, because they are RF=3 tables and written with QUORUM consistency.

But that didn't cover an edge case: if the cluster is single-node, we
cannot query them with QUORUM: it will fail because that requires 2
alive replicas.

The correct way is to query with ONE for single-node clusters and query
with QUORUM for multi-node clusters.

Fixes: #3068