test_audit_categories_invalid(FIX): Don't force wait for cluster to start
The test starts a cluster with an invalid audit configuration, hence the
cluster is doomed to fail starting.
It is assumed that the default value for cluster.force_wait_for_cluster_start
is False but this assumption is apparently not correct.
This simply changes it explicitly to the assumed value (False), so the
test can continue even that the nodes are not going to start.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>