test_audit_empty_settings(Fix): Do not assume audit disabled by default
The test (which is also wrongly named), assumed that audit is disabled
by default. This is not true (at least not to this date).
From reading the test, the true intention is to validate that when audit
is disabled, the 'audit' keyspace is not created unnecessarily.
This test, passes the 'audit : none' configuration instead of empty
settings.

Once passes, a follow up should rename the test to test_audit_none
instead of the current name.

Signed-off-by: Eliran Sinvani <eliransin@scylladb.com>