test(manager): check the tombstone_gc mode during and after restore task
as of the later builds of 3.1 rc, the manager does not alter 'gc_grace_seconds' during
a restore task, but instead it alters the tombstone_gc mode to disabled.
See more at https://github.com/scylladb/scylla-dtest/issues/3108.
As such, I altered the gc_grace_seconds tests to a test that verifies the value of
tombstone_gc during and after the restore.