test(enterprise issue 2852): Terminate a node that has replaced another
In the two new tests, we start a background write stress thread, replace one of the nodes
with a new one, shut the new node down and examine the stress results.
In one test, the new node uses the replaced node's IP, and in the other the new node
uses a new IP address.

See https://github.com/scylladb/scylla-enterprise/issues/2852