dtest_setup: init_default_config: setup all timeout options
Setup the following timeout options, that are not
set at the moment, to complete configuration of all scylla's timeout options:
  counter_write_request_timeout_in_ms
  cas_contention_timeout_in_ms

This is important mostly for debug mode where
all timeouts need to be adjusted to much higher values.

Note that these options are also available for Cassandra,
so it needs no special handling.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>