cdc_test: be more permissive on the format of decorated_Key
before this change, it is required that the decorated_key should
be printed in the format of "{key: pk{KEY}, token:TOKEN}". where
"pk" should be started with a space, while the value of "TOKEN" should
not. we could be more permissive by allowing 0 or more spaces before
the key and token.

See-also https://github.com/scylladb/scylladb/issues/13628

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>