test(manager): restore schema+data in a cluster in a different size
New in manager 3.1.
Added a new test which uses two seperate clusters, each with a different number of nodes.
First, a backup snapshot is created using the first cluster,
and then the test uses said snapshot to restore the schema and data
of the first cluster into the second one.