encryption_at_rest_test: fix running kms test in parallel
since of how the code was configured in case the test
was running inside docker we could end up connecting to
the wrong instance of kms docker.
leading to the following error:

```
>           raise error_class(parsed_response, operation_name)
E           botocore.errorfactory.AlreadyExistsException: An error occurred (AlreadyExistsException)
            when calling the CreateAlias operation: An alias with the name arn:aws:kms:eu-west-2:111122223333:alias/Scylla-test already exists
```

the code was changed to use the specific ip address of the
docker instance, to avoid such cases.