scripts/run_test.sh: pass SCYLLA_PRODUCT to docker correctly
seem like we are not passing `SCYLLA_PRODUCT` into
the docker instance, so we are failing in fetching
enterprise version with CCM code.

in CI we never use this code path, since the pipeline
are the one fetching the version, and it being loaded
before hand from local direcotry, and not fetched
from s3.