auth_roles_test: update keyspace creator permissions
Until recently, the creator of a keyspace was not granted
permissions on all funcitons in this keyspace, so we did not
expect them to appear in the list returned from a LIST ALL
PERMISSIONS statement.
Now that the permissions on functions were implemented, the
expected list of permissions should also contain permissions
on all functions in the created keyspace.