dtest_setup: allow to use local files with scylla-manager tests
Some of the dtests test the behavior of [scylla-manager](https://github.com/scylladb/scylla-manager).
To run them, the caller needs to pass an additional argument to pytest,
`--scylla-manager-package`. This argument points to a place from which
dtest can take the manager and use it for the cluster.

Currently this works on a very specific subset of tests.
It works only when using `version` to describe scylla version
and `--scylla-manager-package` has to point to a remote HTTP resource.

Basically, it works only on the code path that downloads everything
from the internet (both Scylla and manager).

Trying to use it with a local build of Scylla (by passing `--cassandra-dir`),
or using a local folder with `scylla-manager` in it doesn't work at all.
There's no error, the manager just doesn't get downloaded and the test fails
because there's no `sctool`.

This issue was explored in: #3190

The current flow looks like this:

1. User passes the `--scylla-manager-package=<manager link>` flag to `pytest`
2. `dtest_config` sets `self.manager_package` to `<manager link>`
3. `dtest_config` sets the environment variable `SCYLLA_MANAGER_PACKAGE` to `self.manager_package`
4. A `ScyllaCluster` is created, the `version` argument is set, but the `manager` argument is `None`.
5. During `ScyllaCluster` constructor it reads the `SCYLLA_MANAGER_PACKAGE` environment variable
6. It reads the manager link, downloads the package and extracts the contents
7. Then it adds `--scylla-manager=<manager package>` to the `SCYLLA_EXT_OPTS` environment variable
8. Later in `ScyllaCluster.__init__()` it reads the `SCYLLA_EXT_OPTS` environment variable and checks the `--scylla-manager` option.
9. It reads the manager path from it and sets the `manager` variable in `ScyllaCluster` constructor (which was set to `None` previously)

This is a bit convoluted.

This change aims to simplify the flow and enable
the usage of local directories, both for the Scylla
build (`cassandra-dir`) and the manager package
(`--scylla-manager-package`).

The first change is that the responsibility for setting
up the manager is moved from `ccmlib` to `dtest_setup.py`.

Now `DTestSetup` calls `prepare_scylla_manager`, which
looks at the manager link and decides what to do.
* In case of a remote link (starting with `http...`)
  it uses a function from `ccmlib.scylla_repository` to download
  the manager package and extract it to the correct location.
* In other cases it assumes that the manager link
  points to a local directory, and sets the manager
  directory accordingly in its config.

The second change is that we pass information about `manager` using
the `manager` argument of `ScyllaCluster` constructor, not the environment
variable.
The `manager` argument is made precisely to allow specifying where
the manager directory is, there's no need for any environment variables.

Previously the manager was downloaded inside `ccmlib`, so there were
some additional complications, but now `dtest` is responsible for
setting up the manager, so we can just use the `manager` argument
when creating the `ScyllaCluster`.

Note that it's still possible to use the environment variable if needed.
This change modifies only the dtest side of things, `ccmlib` stays as is,
it still reads the `SCYLLA_MANAGER_PACKAGE` and at acts accordingly.
Any outside code that made use of this environment variable should
still work.

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>