bootstrap_test: update expected message in `test_smallest_ip_join_late`
In scylladb/scylladb@297c75c another safety was added to the bootstrap
procedure - the bootstrapping node waits for another node to show up in
gossip before waiting for schema agreement, or times out with an error
if it doesn't after some time.

This causes `test_smallest_ip_join_late` to fail because it expected
another error to happen on a later step of the boot procedure.

Modify the regex in `start_failure_msgs` to add the new error message as
an alternative.

Fixes #3203