cdc_test.py: make wait_other_notice the default
cdc_test.py has its own cluster-creating utility functions, and they
have a wait_other_notice parameter. Creating a cluster with this
parameter False makes little sense - it means we don't wait for nodes
to be aware of each other and be usable from each other, and if the
test starts immediately after the cluster is constructed, the cluster
may not yet be usable because not all replicas are usable by
coordinators.

So it makes more sense for the wait_other_notice parameter to default
to True.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>