conftest: add the cluster_options marker
"cluster_options" accepts kwargs arguments, each keyword will be
considered as the option name, and its value the value of the option.
these kwargs will be used to update the cluster_options before the
cluster is created.

this change would allow us to customize multiple options for a
test or a test suite in a simpler way.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>