disable uuid_sstable_identifier_enabled for some tests
these tests are identified to be not able to support
uuid_sstable_identifier_enabled=True. as they are using either
sstableloader or sstabledump. these tools are based on Cassandra 3.x,
and not support the uuid-based sstable identifiers. before we switch
over to alternatives of these tool or update them to support this
feature, let's just disable this option for these tests.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>