system_tables_test: convert 0/1 to false/true to match values in scylla.yaml
Configuration options of type `restriction mode` are specified using `false`,
`warn` and `true` in scylla.yaml, but in system.config the corresponding values
are `0`, `warn` and `1`. Convert `0` -> `false` and `1` -> `true` for the check to pass.

The test test_content_source_config  started failing after adding a new configuration option:
`strict_is_not_null_in_views`. This option is set to `true` in scylla.yaml,
but the value read from scylla.config was `1`, which caused the check to fail.

Fixes: https://github.com/scylladb/scylla-dtest/issues/3225

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>