migration_test: disable uuid_sstable_identifier_enabled for some test
the test of "test_migrate_sstable_with_large_row_number" uses
sstableloader to load the backup of sstables, but sstableloader
does not support uuid-based sstable identifiers. before we switch
over to alternatives of these tool or update them to support this
feature, let's just disable this option for this test.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>