fix(materialized_views_test): remove invalid IS NOT NULL restriction
In materialized views the IS NOT NULL restriction
works only on columns which belong to the view's
primary key.
In the past we used allow IS NOT NULL restrictions
on other columns by mistake, but they were just silently ignored.

This was a bug, it shouldn't be allowed, and was fixed in
https://github.com/scylladb/scylladb/pull/13013.

One dtest had such an invalid view definition,
and started failing because Scylla now rejects
its CREATE MATERIALIZED VIEW statement.

Let's remove the extraneous IS NOT NULL condition
to make the statement valid again.

Fixes: https://github.com/scylladb/scylla-dtest/issues/3228

Signed-off-by: Jan Ciolek <jan.ciolek@scylladb.com>