tools/files: get_list_of_sstables: look for Data.db files by default
I'm really not sure why the default is the Statistics.db files.
Most callers are interested only in the sstables count
so they don't care what of the sstable components is returned
while other that are interested in the sstables either
specify the suffix or would be interested typically in the data file.

For example, test_cleanup_space_amplification checks the file size
to verify space amplification during cleanup, and checking the
statistics component size is completely wrong.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>