test_cleanup_space_amplification: measure data file blocks
Scylla sets the logical file size upwards on purpose
and at the end it truncates it to the real length.
During cleanup we may see sstables files during their write
(especially in debug mode), so the logical length is not
a good measure.  We shuold look at the used blocks instead
to verify space amplification.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>